Non integer particle count spawning

+++ common/src/main/java/net/combatspells/api/spell/ParticleBatch.java
@@ -10,13 +10,13 @@ public class ParticleBatch {
     public enum Origin {
         FEET, CENTER, HANDS
     }
-    public int count = 1;
+    public float count = 1;
     public float min_speed = 0;
     public float max_speed = 1;
 
     public ParticleBatch() { }
 
-    public ParticleBatch(String particle_id, Shape shape, Origin origin, int count, float min_speed, float max_speed) {
+    public ParticleBatch(String particle_id, Shape shape, Origin origin, float count, float min_speed, float max_speed) {
         this.particle_id = particle_id;
         this.shape = shape;
         this.origin = origin;
