+++ common/src/main/java/net/combatspells/network/Packets.java
@@ -63,7 +63,7 @@ public class Packets {
             buffer.writeString(batch.particle_id);
             buffer.writeInt(batch.shape.ordinal());
             buffer.writeInt(batch.origin.ordinal());
-            buffer.writeInt(batch.count);
+            buffer.writeFloat(batch.count);
             buffer.writeFloat(batch.min_speed);
             buffer.writeFloat(batch.max_speed);
         }
@@ -73,7 +73,7 @@ public class Packets {
                     buffer.readString(),
                     ParticleBatch.Shape.values()[buffer.readInt()],
                     ParticleBatch.Origin.values()[buffer.readInt()],
-                    buffer.readInt(),
+                    buffer.readFloat(),
                     buffer.readFloat(),
                     buffer.readFloat()
             );
